Fix potential namespace collsion in ftplugin/ruby.vim
On win32 installs with +ruby, ftplugin/ruby.vim makes a call to
VIM::command to set the s:ruby_paths variable. In some circumstances,
this code can run when under a different ruby module, and if that ruby
module has a VIM module defined, there is a namespace clash. The fix is
to change the call to be ::VIM::command, which works fine under all
cases. The conflict occurs with Command-T in particular when opening
a .rb file from that tool.