Attempt to fix problem with incorrect heredoc highlighting
The \_s* matcher was being used to correctly highlight constructs like
"class << self" and "Foo::<<", even when they're on separate lines and
oddly spaced.

This commit separates the patterns for these groups from the group with
the closing brackets, [)}...], since in that case, it shouldn't be
necessary to match the << on a new line.