Special rules for `def` as an indent keyword
See https://github.com/vim-ruby/vim-ruby/issues/190

Ruby 2.1 makes method definitions return the method name as a symbol.
This means that you could prefix the `def` line with a `private`
and it would make the method private upon definition.

Technically, any method call could be placed before the `def`, but
that's no different than before. The private/protected use case seems
like the most likely one, so this is what we'll handle for now. If it
gets popular to use different class-level methods for method
definitions, we could put a \k\+ instead.