Fix blocks in the middle of dot continuations
The deleted MSL case seems like it has the purpose of only taking the
closest line of continuations, and ignoring the rest so it doesn't get
confused by the (potentially unindented) previous line. For example:

  one.
    two.
    three do
      puts "OK"
    end

In this case, the MSL search will find "two." and immediately stop
there, taking +2 spaces from "two.". If it did continue, it'd find
"one." as well, and respect that one.

However, if the "two." wasn't there...

  one.
    three do
    puts "OK"
  end

This changes things simply because the "one." line becomes the MSL, and
the "three" is indented with respect to it.