Don't highlight `.?c` as a "character" literal.
`.?`, the "safe navigation operator", has been accepted into Ruby and
presently conflicts with the rule for highlighting `?c` character
literals. This patch remedies the issue by adding `\.` to the class of
preceding characters which refute such a match.

I refrained from adding `.?` to `rubyOperator` only for symmetry with
the absence of `.` therein, but `.?` is sufficiently "weird" as to be
worth considering making an exception for; it's the sort of thing that
really ought to stand out, in my estimation.

That `::` can be used for method invocation seems to have been largely
forgotten, but I suspect `::?` will eventually find its way in, at which
time another patch very much like this one will be necessitated.