Fix bug with blocks in argument lists
A "non-bracket" continuation seems to be too wide of a net. Introducing
a "dot" continuation helps to fix the issue, since it seems like method
chains in particular have some specific behaviour.

Instead of copying the regexes around, it might make sense to create a
"Continuation" object or something that holds various properties of the
given line -- does it end in a comma, dot, bracket, etc. This might be a
simple refactoring that would make the code slightly more readable. A
next step might be to add additional information to this object, like the
columns of unclosed brackets, that would help with fixing some
long-standing issues with pairs of brackets.

Let's see if this ever gets done.