Handle backtick continuations
While they generally seemed to work in most cases, there wasn't really
any explicit handling of this sort of continuation, and in the case
described in the added test, the presence of a `.` and `#` confused
pattern-matching somehow in the case of searching for an MSL. The
indentation for the line after the `\` ended up falling back to the
final `return` where we would just return the previous line's indent.

Unfortunately, I'm still fuzzy on the details of how this was
possible and why this isn't a problem in *all* cases, not just ones with
a `.` and a `#` following. It's likely that future issues will provide
more information to triangulate the problem.