Fix issue with default block parameters
The block pipe is considered a line continuation. By expanding `\W` into
`[^[:alnum:]]` and adding the `:` to it, the regex only matches pipes
that don't have `:` before them. Unfortunately, a space messes this up,
so we add a `:\@<!` to check if there wasn't a space before said
alphanumeric character.

Just adding extra whitespace wouldn't really work, since the standard
pipe continuation breaks. `foo | bar` is a perfectly reasonable "bitwise
or" operator, which has a word character, followed by a space...

I can't currently think of a scenario where this would be a problem,
even if it does complicate the regexes a bit.