Try to fix a heredoc highlighting issue
The problem is that the following code highlights the first `EOS` as a
constant:

  def foo()
    <<-EOS
    EOS
  end

This seems to be related to the `\_s` in the negative lookbehind
pattern. The "<<" is considered the start of a heredoc when it's not
prefixed by something that ends with ")" (which sort of makes sense,
might be a shift operator), but in a multiline case, this seems
incorrect. Changing the whitespaciness of the `\_s` seems to fix the
issue.

This has already been done for *some* heredoc patterns, but not all, it
seems.