Optimize the "empty line inside a string" case
In a large heredoc, calling s:PrevNonBlankNonString can be super
expensive, because it goes over every single line, checking its syntax
groups. But we don't need the knowledge of the real "previous non-string
line" in that situation, we only use it to check if it's different than
the one `prevnonblank()` gives us.