Try to fix a module loading bug
If searching with the "r" flag, the outermost match will be found, but
this seems to find the wrong end match. For example:

    module Foo
    end

    module Test
    end

    Foo::

In this case, completing `Foo::` will include `module`, because the
found `end` is the one closing `Test`, not the one closing `Foo`.

By removing the "r" flag, the module seems to be detected correctly, but
hard to say if this will have any side effects.