Improve backslash escaping in double quoted literals
All escaped characters are now highlighted as such even if they have
no special meaning.  This is a more accurate representation and appears
to be an approach used extensively but not exclusively in other syntax
definitions.  It also simplifies the implementation considerably.

Fixes a few odd long-standing oversights like \" and \$.

This further addresses #286.