Fix conditional highlighting after redundant line continuations
Conditional and loop modifiers always appear after required line
continuations so match the redundant ones and allow the general forms to
match as normal.

Adds a user configuration variable, ruby_line_continuation_error, to
highlight useless line continuations as errors.

Addresses the syntax highlighting component of #340.

Because you have to be more than a little mad to be doing this in the
first place.