Start all specs with default indenting globals
Half of the spec files would fail when run independently, but pass when
running the whole folder. This is due to some files modifying global
variables, which then affect other files.

Now, all specs run with globals set to default values, which need to be
explicitly set to something else if required by a test.

Existing tests that implicitly expected non-default values have been
given those non-default values explicitly. The other option was to
update those tests to use the defaults, but I didn't want to mess with
the expected values during refactoring.