Remove broken `s:IsInString` check
The variable `line` doesn't really exist, but we've never gotten an
error report, and I can't come up with a valid ruby syntax that actually
hits this clause. The check looks at the end of the line and whether
it's a string or comment, but we already do that in
`s:MultilineStringOrLineComment`.

The issue was discovered on this text:

  Indent tests failed:
  testdir/html.fail
  indent (this) line;

So, a colon on the first line, some text that doesn't end in a
continuation character, and the last line errors out. But I couldn't
come up with a syntactically valid ruby combination that would hit it.
So I removed the check completely.